requirements.txt
streamlit~=1.43.1
google-generativeai
PyPDF2~=3.0.1
python-docx
pandas
openpyxl
sentence-transformers
faiss-cpu
python-dotenv~=1.0.0
numpy
langchain-community~=0.3.19
chromadb
langchain_google_genai
protobuf~=3.20.3
langchain~=0.3.20
langchain-google-genai~=2.0.10
langchain-openai~=0.3.8
requests~=2.32.3